+++ js/util.js
@@ -1,0 +1,17 @@
+export const getRandomInt = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(Math.random() * (upper - lower + 1) + lower);
+};
+
+export const getRandomArrayElement = (items) => items[getRandomInt(0, items.lenght - 1)];
+
+export const generateCommentId = () => {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
